handle set episode rating

+++ core/data/remote/src/main/java/com/chocolatecake/remote/request/RatingEpisodeDetailsRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class RatingEpisodeDetailsRequest(
     @SerializedName("value")
-    val value: Int
+    val value: Float
 )
